Video-detection-and-segmentation/requirements.txt
ultralytics
ffmpeg-python
